Create Second HTML File
This File contains codes which gets you introduced to the concept of Hyperlinks in detail. You will learn how to fetch a file inside a Folder of your computer. 
Apart from this the End Section contains the concepts of Image and how to increase or decrease the size of image on the Webpage.